+++ app/src/main/java/ch/mcfx/urs/data/local/AppDatabase.kt
@@ -20,7 +20,7 @@ import androidx.room.TypeConverters
         ListEntity::class,
         ListItemEntity::class,
     ],
-    version = 7,
+    version = 8,
     exportSchema = false,
 )
 @TypeConverters(Converters::class)